bstr 1.0 -> 1.4.0, time 0.3 -> =0.3.17, git-config 0.12 -> gix-config 0.16.3
Summary:
bstr introduced a change in 1.3.0 that will break `AsRef` deduction.

See: https://github.com/BurntSushi/bstr/issues/149

`git-config` is affected but the patch fixing this is only available in the renamed crate `gix-config`.

This causes a `time` to be upgraded as well. However, in `time` 0.3.18 the crate removed the ability to use `unsound_local_offset` to opt out soundness check we rely on. Thus we are pinning `time` to 0.3.17 for now.

This change also caused windows 0.40.0 to be removed but it pulled in windows 0.43.0 because of `git-sec`. Upstream has upgraded the crate but they haven't made a release yet.

allow-large-files

Reviewed By: zertosh

Differential Revision: D45285687

fbshipit-source-id: 88b22d315e92ecced7297a1e672c63ee14437ce3