Stop using preg_match for our unit tests
Summary:
For expectf or expectregex test files, we use preg_match to verify the
test output. The problem with this is that preg_match can fail to
compile the regex if it's too complicated. Our expect files tend to
have a lot of literal text and that especially triggers the
failure. It happens even more easily when we run with multiple
requests because the size of the regex is increased. When the
preg_match failure happens, we "skip" the test. I'd argue this is very
bad behavior, as it means the test you're intending to run may never
actually run.

The syntax was use for expectf files is relatively straight-forward
and we don't really need to use preg_match for it. Without too much
difficultly we can write a custom expectf parser which provides the
matching behavior. This is what I've done. While we're at it, with a
few more additions, we can extend the expectf format to support
everything we need from expectregex. Thus we don't need preg_match at
all.

Convert the expectregex files over to use the new expectf format. Add
a new test runner rule that you can only use an expectf file if you're
actually using wildcards. If not, you have to use a plain expect file
instead. Convert all of the affected tests over to expect.

Rather than relying on a copied matching pattern to deal with multiple
count output, add the ability to set an environment variable to place
a delimeter in between run. The test runner can use this and then
split the output according to that delimeter. Each split component of
the output can be matched separately.

Cleaning up all of this exposed a few tests which weren't running in
--retranslate-all, and now that they do, fail. Fix those (mainly by
adding nomultireq).

Finally, remove all disabled tests. Many of them have been broken
forever and if we're not actively fixing them, we don't need to keep
them around.

Reviewed By: aorenste

Differential Revision: D45296962

fbshipit-source-id: b66e4a4d143a5e3bf9e3d1084f1ea924eeac7e28