Improve crash detection and reporting in hphp/test/run.php
Summary:
hphp/test/run.php will fail a test if the tests writes anything to
stderr. In the process it will print out the contents of stderr. This
is the only explicit crash detection. However, this doesn't accomplish
much because the test runner redirects HHVM's stderr into stdout. It
must do this because a HHVM test might write arbitrary things to
stderr which aren't indicative of a test failure. However this
redirection means that if HHVM, we only get any notification because
the wrapping /bin/sh invocation prints out the command-line! The
actual crash report from HHVM is mixed in with stdout and not
displayed.

Attempt to improve this situation. First of all, if we detect a crash,
display stdout in addition to stderr. Since stdout will have the test
output in it as well, attempt to look for the typical signs of a crash
dump and only print out that portion. If that fails, just print out
the last 100 lines of stdout. Finally, infer a crash not only if
stderr is printed to, but if HHVM returns any exit code other than 0
or 1. 0 means success, and 1 typically means a "failure" like a fatal,
which is considered a normal execution of the test.

A few tests return a non-0 and non-1 exit code even when they
succeed. To support these, add support for an optional exit_code file,
containing the exit code the test is expected to return.

Reviewed By: oulgen

Differential Revision: D45029944

fbshipit-source-id: 0bfbff9d67a67692893eb6d4408b2a30dc6573ad