track reason for cancellation
Summary:
A user reported:
> `Files have changed! Please re-run hh.` I keep getting that at the end of a very long run (20+ minutes), even when I'm not touching the IDE

The user's log revealed five instances of this message - two because EntMutationBuilder.php had changed on disk, two because of edits in the IDE, and one because of file open/close in the IDE.

I think that the "Files have changed" message must include an example file+timestamp in case of file-change, and an explanation about the IDE action in case of IDE change. Otherwise users will find the message too perplexing.

This diff adds half the plumbing. It changes it so that if an interrupt-handler returns `MultiThreadedCall.Cancel` (the only means by which cancellation can happen), then through the ocaml type system it must now also provide a reason. It plumbs that reason all the way through into the return type of `Typing_check_service.go_with_interrupt` (whether through distc interrupt handler or MultiWorker interrupt handler), where it's consumed by `ServerTypeCheck.do_type_checking`. The next diff in the stack will take it from there all the way into the errors-file, where it will be printed by `ClientCheckStatus.go_streaming`.

I had a design choice for this diff:
1. In the interrupt handlers in serverMain.ml, store the cancellation reason in `env`, and rely on runtime invariants that it will hopefully be set in all the right places
2. In the interrupt handlers in serverMain.ml, return the cancellation reason, and change the ocaml typing rules of MultiThreadedCall to ensure that the reason gets plumbed right up to its caller.

In this diff I went for option 2 because I believe that runtime invariants around `env` and ServerTypeCheck are too fragile. (Spoiler: although I'm able to use the ocaml type system to guarantee reasons get propagated into `ServerTypeCheck.do_type_checking`, I think I'll have to resort to runtime invariants for its caller `type_check_core`, nor its caller `type_check` where the error-file is done. More on that in the next diff. Even so, it's still nice that I'll have been able to isolate the runtime invariants to just this group of three functions.)

## How to read this diff

1. Look at `multiThreadedCall.mli` where the ocaml type system shows that a cancellation reason must be provided
2. Look at `multiThreadedCall.ml` to see where MultiWorker plumbs that reason through to its callers
3. Look at `typing_check_service.ml` to see where distc plumbs that reason through to its callers
4. The rest of the diff is just more plumbing to carry the reason through to where it belongs

Reviewed By: CatherineGasnier

Differential Revision: D45219650

fbshipit-source-id: 2e4e84188c571897362c5e81bfdb6baad892c221