Use unknown type nothing for closed shapes
Summary:
Following the change to shape introduced in D45001642, where the type of unknown fields in open shapes is specified explicitly, we now use this type to express *closed* shapes as open shapes in which the unknown fields have type `nothing`. So, for example:
```
shape('a' => int) == shape('a' => int, _ => nothing)
```
This simplifies slightly the representation and logic associated with shapes.

Reviewed By: vsiles

Differential Revision: D45231796

fbshipit-source-id: 795ed08ce304b71281e29bc16ff1346a43adad83