Nested explicit construction of dictionary should not have like type
Summary:
Sometimes we were inferring like types in collections even if no dynamics were involved. For example:
```
$f = dict<string, dict<string, int>>['bar' => dict['baz' => 42]];
```
We infer a type of `dict<string, ~dict<string,int>>`, because an expected type with implicit like (here, `~dict<string,int>`) is passed down to checking of the subexpression `dict['baz' => 42]` and the like does not get stripped when producing an inferred type.

We change the logic of checking collections to strip this like. Note that we need to do this underneath intersections e.g. we might have an expected type `~int & arraykey`.

Differential Revision: D45326925

fbshipit-source-id: b6ac0dc3442e2769583fcc8e81787a038d898d04