print reason for interruption
Summary:
The user scenario, discussed in the previous diff:
> `Files have changed! Please re-run hh.` I keep getting that at the end of a very long run (20+ minutes), even when I'm not touching the IDE

What happened was that `Typing_check_service.process_in_parallel` got interrupted because of an interrupt handler (either watchman file-change, or priority channel, or persistent connection) which told it to `MultiThreadedCall.Cancel` the typecheck. The previous diff added a reason for the interrupt+cancel `MultiThreadCall.Cancel reason` and threaded it through to the return of `process_in_parallel`.

This diff takes that reason and exposes it to the user.

## How to review this diff
1. Read serverEnv.mli for the new field `why_needs_server_type_check`
2. Read serverTypeCheck.ml for how that field is read+written
3. Read ServerProgress.ml for where it gets written into the errors-file
4. Read clientCheckStatus.ml for how it gets read from the errors-file

I stored it as `(string * string)` rather than a named record type. That's a shame. I did that because it needs to belong in MultiThreadedProc.ml, ServerProgress.ml, and ServerEnv.ml. The first two are very low-level libraries, so I'd need an even lower-level library to define the type in. I didn't really think it was appropriate to put as high-level a concept as "why was the typecheck interrupted" into a low level library.

The pure way to do this would be to parameterize ServerEnv and ServerProgress upon the type of cancellation message. But that would be a lot of code complexity and hard-to-read-ness, for something that totally doesn't deserve it.

Reviewed By: CatherineGasnier

Differential Revision: D45253908

fbshipit-source-id: d22195402cf9c233fd2f55337b7d5a40904bf5e6