don't interrupt+cancel on Open/Close
Summary:
A user reported:
> `Files have changed! Please re-run hh.` I keep getting that at the end of a very long run (20+ minutes), even when I'm not touching the IDE

The user's log revealed that they were opening/closing files in the IDE. (I guess they didn't consider this "touching"). As things stand, opening/closing/editing has the effect for hh_server of changing the source of truth for a file's contents. And this triggers any ongoing check to be cancelled and a new one started, exactly the same as happens when we discover through watchman that file-content on disk has changed.

(Well, not quite the same. For watchman changes we are only ever notified after the fact, and so there's a race possibility where the ongoing typecheck might have read the new file contents before being aware of it; for Open/Close/Edit, we actually delay handling the Open/Close/Edit until the ongoing typecheck has been brought to a halt).

What this diff does is make it so that Open/Close don't count as changing the file's contents. (well, so long as whatever VSCode opened/closed the file with is the same as what it was on disk). This way, users can happily browse files while a typecheck is ongoing.

------------------

Note: once I made this change, I discovered there's a further irritation lurking.
```
           IDE edits can come in quick succession and be immediately followed
           by time sensitivie queries (like autocomplete). There is a constant cost
           to stopping and resuming the global typechecking jobs, which leads to
           flaky experience. Here we set the flag [recheck_restart_is_needed] to [false] for
           Edit, meaning that after we have finished handling the edit then
           [ServerMain.persistent_client_interrupt_handler] will stop the natural
           full check from taking place (by setting [env.full_check_status = Full_check_needed]).
           The flag only has effect in the "false" direction which stops the full check
           from taking place; setting it to "true" will not force an already-stopped full check to resume.

           So what does cause a full check to resume? There are two heuristics, both of them crummy,
           aimed at making a decentish user experience where (1) we do not resume so aggressively
           that we pay the heavy cost of starting+stopping, (2) the user is rarely too perplexed
           at why things do not seem to be proceeding.
           * In [ServerMain.watchman_interrupt_handler], if a file on disk is modified, then
             the typecheck will resume.
           * In [ServerMain.recheck_until_no_changes_left], if there is a pending "hh" command
             still connected via RPC, and it has been 5.0s or more since the last Edit, then the typecheck
             will resume. (Note that this does not apply to streaming-errors in [ClientCheckStatus.go_streaming],
             since these are not done via RPC to hh_server). *)
```
I'll address this in the following diff.

Reviewed By: CatherineGasnier

Differential Revision: D45305006

fbshipit-source-id: 459b743f90ec9885916ab94e108271bce9869fbe