Cache large localization results
Summary:
In certain cases, we can cache the output of localization and get a performance boost.

The conditions for caching are:
1. No errors occurred
2. No cycles were detected (those are not consistently reported as errors)
3. The expansion happens under regular assumptions (newtypes are visible if they are in the current file)
4. No global type parameters were created (e.g., `C::TAbstract` when we bogusly create abstract type constants); if we memoized in this case we would fail to reconstruct the bounds on the global parameters when there are some
5. The expanded type's reasons do not depend on the input decl type

To satisfy condition 2, we add tracking for whether a cycle was encountered in the expand environment.
To satisfy condition 5, we limit the memoization to shape types expanded from type constants or type aliases. For those, most of the reasons (except the toplevel one that we patch) come from the type constant/alias definition site and do not depend on the input decl type.

We also reset the cache as soon as the current file or module changes; these affect how newtypes are expanded.

This extensive list of conditions under which localization behaves in a cacheable manner may look concerning. Nonetheless, I reckon it is valuable to have it all explicitly laid out in code so that future localization cleanups have an up-to-date reference point.

Reviewed By: vsiles

Differential Revision: D44054734

fbshipit-source-id: ec68628616e8d044813a103c89b5995b77fb31b0