InMemoryView: do not warm sha1 cache for large files
Summary:
Computing sha1 for very large files can be extremely expensive, and in the case
where that file keeps beeing written to, Watchman can end-up burning CPU cycles
and all of the cores of the system for the cache to be invalid by the time the
data is needed.

This introduces a new config that allows the cache warmer to skip over large
files (>10MB) to avoid falling into this pathological case. The downside is
that if the sha1 is needed it'll be computed on the fly and then stored in the
cache.

Reviewed By: fanzeyi

Differential Revision: D45377529

fbshipit-source-id: 69c9b0cde1bb72978af22baf92414375d54a47da