make the module containing mock fields pub so it shows up in rustdocs
Summary:
the rustdocs for mocks can be a bit confusing because there's no documentation for any of the individual mock fields. this is because the module in which the mock field objects are defined is private.

to expose them in rustdoc, just `pub` that module. each of the mock field objects have private fields, so downstream crates will still be unable to instantiate these mock fields on their own (and this makes `unimplemented` `pub(crate)` to ensure that as well).

the link to some mock fields is still broken, e.g. `FacebookService`, but that's because of https://github.com/rust-lang/docs.rs/issues/1588#issuecomment-1464958351.

Reviewed By: thepacketgeek

Differential Revision: D45406779

fbshipit-source-id: 91f14ef2e022857969540c82d2482e81bda1421c