Recursive mkdir isn't multi-process safe
Summary:
A recursive mkdir is supposed to create any intermediate directories
as necessary, and only error if the final directory already exists. To
accomplish this, mkdir_recursive calls ::access to check if the
directory exists before calling ::mkdir. If ::mkdir returns an error,
the whole thing fails. This doesn't work if there's multiple
concurrent thread/processes calling mkdir with the same directory
prefix. The ::access check may signal the directory doesn't exist, it
gets created in the meantime, then the ::mkdir call fails.

Fix this by skipping the ::access check in the loop and just calling
::mkdir. If it fails due to EEXIST, just ignore it and
continue. EEXIST is only an error on the last directory.

Reviewed By: mdko

Differential Revision: D45407098

fbshipit-source-id: 82de656ae4a2ecc85af0cae8a8c153114d7daf92