fix Exit_status -> Exception dependency
Summary:
Currently, Exception.ml has a dependency on exit_status for no good reason.

It's doing this to pretty-print the Exit_status.Exit_with exception. But the correct ocaml way, as used in many other places in the code, is to do this is to register a printer for this exception type.

(I'm doing this because I want to add richness to some of the Exit_status exceptions, in particular so they include finale_data, to allow for better telemetry).

Reviewed By: bobrenjc93

Differential Revision: D45408062

fbshipit-source-id: 1b9dec79213df173f3a9393d89256b34264cc581