CLIENT_BAD_EXIT shows reason for server_hung_up
Summary:
I'm trying to set up an OKR for how the server succeeded or failed in "hh check", and how long it took. It might fail with CLIENT_CHECK_BAD_EXIT. The reason for this is 60% server_hung_up, 30% ctrl+c, and 10% others. https://fburl.com/scuba/hh_server_events/j57d64a6
{F964109901}

To make the OKR richer, I'd like to include information about *why* the server hung up. That's already available in a completely different event from the client, SERVER_HUNG_UP, whenever ClientConnect.connect fails.
{F964113485}

But that's not a great place for it! (1) Scuba can't do joins on two different events, (2) it doesn't even happen for streaming errors. To make the OKR straightforward, I want to carry the data through to the CLIENT_CHECK_BAD_EXIT event for both streaming and regular errors. That's what this diff achieves.

Reviewed By: bobrenjc93

Differential Revision: D45408063

fbshipit-source-id: dc4faaba340d49ccd775f759e1d47a2322ca5806