tracking down empty paths
Summary:
I'm continuing my mission to remove the need for `~default_path` when we convert a `Pos.absolute` to an Lsp position.

The issue is that sometimes we see `Pos.absolute` with empty filename. I'd added telemetry to see how prevalent this is, and where. https://fburl.com/scuba/hh_server_events/is3x1czl - the answer is
* we get it from do_server_diagnostics, which calls hack_pos_to_lsp_location on each constituent message of each error
* some of these constituent messages have empty filename

This diff adjusts the error-processing path to explicitly acknowledge and handle this case. Once this has landed, telemetry will hopefully show that there are no other cases with empty filenames.

Reviewed By: Wilfred

Differential Revision: D45346064

fbshipit-source-id: 0a8910d829fce5069fef20e713644408f469331a