track whether diagnostics came from errors-file or clientIdeDaemon
Summary:
In a future diff I want to add temporary validation that the diagnostics we publish out of clientIdeDaemon are the same as what's reported in the errors-file.

I therefore need to know whether the errors we've published for a given file came from clientIdeDaemon or from errors-file. That's what this diff does.

More generally, it just kind of feels nice (e.g. for debugging) to be able to track where the errors came from. So even though the need is temporary, I think this is a worthwhile thing to add.

Reviewed By: CatherineGasnier

Differential Revision: D45155386

fbshipit-source-id: ce33cd6329d01b29d4e52ae01f6107bba1afe939