clientIdeDaemon returns errors after codeAction
Summary:
An earlier diff in the stack had made it so that updated errors go from `clientLsp->VSCode` for didClose+codeAction. This diff puts in the `clientIdeDaemon->clientLsp` bit for codeAction.

Reminder: this docblock in clientIdeMessage explains why errors are being sent in codeAction:
```
  | Code_action :
      document * Ide_api_types.range
      -> (Lsp.CodeAction.command_or_action list * Errors.t option) t
      (** Handles "textDocument/codeActions" LSP messages.

      Also, we take this as a handy opportunity to send updated [Errors.t]
      to clientLsp so it can publish them. We return [None] if the TAST+errors
      have not changed since they were last sent to clientLsp (through didOpen,
      didChange, didClose or an earlier codeAction), and [Some] if they have
      been changed.

      Why send errors in code-action? It is for the scenario where file A.PHP
      is open, file B.PHP changes on disk (e.g. due to user saving it), which
      invalidates the TAST+errors of A.PHP, and we might therefore need to send
      updated errors. For efficiency we chose to do this lazily when the user
      interacts with A.PHP or switches tab to it, rather than doing every single
      open file up-front. We could have sent updated errors on every single action,
      e.g. hover and go-to-def. But settled for now on only doing it for codeAction
      because this is called so frequently by VSCode (when you switch tab, and every
      time the caret moves) in the hope that this will be a good balance of simple
      code and decent experience. *)
```

Per the docblock, we have to keep track of what errors have last been published from clientIdeDaemon by didOpen+didChange, so that codeAction can avoid re-publishing errors if they haven't change. What's the best way to do this? I first wrote it using immutable data D45111175, but that took 143 lines of not-very-nice code to achieve. This diff instead does it with mutable data and accomplishes it in only 55 lines of nicer-to-read code.

I've used `phys_equal` to determine whether the errors that were last published from clientIdeDaemon are identical to what the current errors are. That felt like a nice (robust, non-invasive) way to keep track.

Why did I go to the trouble of avoiding republishing in case errors are the same? This docblock explains:
```
      match !published_errors_ref with
      | Some published_errors when phys_equal published_errors errors ->
        (* If the previous errors we returned are physically equal, that is an indication
           that the entry\'s TAST+errors has not been recomputed since last we returned errors
           back to clientLsp, so no need to do anything.
           And we actively WANT to do nothing in this case, since codeAction is called so frequently --
           e.g. every time the caret moves -- and we would not want errors to be republished that
           frequently. *)
        None
```

Reviewed By: Wilfred

Differential Revision: D45128709

fbshipit-source-id: ba51f0cf0dfb2d27c7dadf57fc05f0a20a3720cc