Schedule work better for test runner workers
Summary:
Right now the test runner just divides all the tests into N equal size
buckets for N workers. Since the list is alphabetical, this means the
same worker will run a lot of similarly named tests. This leads to
inefficiences, as one worker can get a lot of slow running tests. All
of the other workers will finish, but we'll be stuck waiting for all
of the slower tests to finish.

Instead of pre-dividing work, use a work queue approach. Instead of
being told what tests to run, each worker just pulls the "next" test
off of the work queue. So, even if one worker is bogged down running a
test, the rest of the tests can get picked up in the meantime.

Of course, being PHP/Hack, doing such a work queue is not
straight-forward, especially since the workers fork. Instead of having
a literal work queue, each worker has a copy of the entire test
list. So, finding the next test is just incrementing an index
atomically. Since the workers are different processes, we need some
IPC mechanism for this. The simplest way is to store the index in a
file and use flock to ensure it can only be accessed concurrently by a
single worker at a time. To pull the next test, the worker opens the
file, locks it, reads the index, writes the incremented index, then
closes the file (releasing the lock).

Reviewed By: oulgen

Differential Revision: D45411038

fbshipit-source-id: 08a70a53d32e86a6f2748d4a4b8da34a3ff0918c