Filter out files not in the current deployment during repo build time
Summary:
This diff adds support for when an active deployment is specified, the files that do not belong to packages specified in the active deployment to be removed from the build.

Note that one package can be in multiple deployments but each module must only be in at most one package.

Reviewed By: edwinsmith

Differential Revision: D44852384

fbshipit-source-id: 1bae90881cdf834851145ed58c6e08a49dac27e9