ServerInitCommon rename
Summary:
[This is part of an 19-diff stack whose purpose is to remove the redundant Errors.Parsing and Errors.Decl phases]

The functions in ServerInitCommon use our conventional names for things "indexing, parsing, naming, type_checking". But I've come to believe these names are hindering our collective understanding...
* "parsing" only uses direct-decl-parser as a means to achieve its actual end of updating the forward naming table
* "naming" is a horrible thing that updates the global mutable reverse naming table and generates "duplicate name" errors
* "type_checking" actually only defers type-checking, rather than doing it, in the common case.

This diff therefore renames the functions to be more descriptive. I hope that this will make it easier for people to review the changes I plan to make in ServerLazyInit.ml...

In renaming the functions, I encountered this one. The name wasn't quite descriptive enough, and in trying to figure out what it does, I came to convince myself that it never actually even uses half of its functionality. I therefore added a `HackEventLogger.invariant_violation_bug` to confirm.
```
(** This function ensures the naming-table is ready for us to do "naming" on the dirty files
("parsing_files"), i.e. check all the symbols in them, add them to the naming table, and report
errors if any of the dirty files had duplicate names.
*)
let naming_from_saved_state ...
```

Note: a change I'm considering for the future is I think it's weird how `serverLazyInit.post_saved_state_initialization` builds up a naming-table-delta in `env.naming_table`, and then several functions take in an `env` and return an `env` when they operate upon that delta, and then it merges the "Backed" loaded-from-saved-state with that delta to produce the final `env.naming_table`. I think it'd be easier to read if the naming-table-delta were a local variable, and `env.naming_table` is the Backed thing right from the start. But I'll save that for future.

Reviewed By: bobrenjc93, CatherineGasnier, rjbailey

Differential Revision: D45005257

fbshipit-source-id: ae62607fa8bcaf8e61acab92d552484a6bf58a39