Decl_service doesn't produce Errors.Decl
Summary:
[This is part of an 19-diff stack whose purpose is to remove the redundant Errors.Parsing and Errors.Decl phases]

As discussed in the previous diff, I proved that Decl_service's call to `Decl.make_env` never produces any errors. This code is straightforward cleanup as a consequence.

Decl_service is only ever called by serverEagerInit. (not to be confused with Decl_redecl_service which is the subject of the following diff).

Reviewed By: bobrenjc93

Differential Revision: D44901778

fbshipit-source-id: 57f41fe2319c60bfdee82652491f35702bb539f4