Decl_redecl_service doesn't produce Errors.Decl
Summary:
[This is part of an 19-diff stack whose purpose is to remove the redundant Errors.Parsing and Errors.Decl phases]

As discussed in the previous-but-one diff, I proved that Decl_redecl_service's call to `Decl.make_env` never produces any errors. This code is straightforward cleanup as a consequence.

Decl_redecl_service is only ever called by serverTypeCheck. (not to be confused with Decl_service which was the subject of the previous diff).

Reviewed By: bobrenjc93

Differential Revision: D44901776

fbshipit-source-id: 76ebb1656c5a46df25794ed11ec0b80e81a3dc78