Fix string regex for stack trace
Summary: The string no longer properly reflects how the functions are scheduled. co_funcA indeed appears in a second trace, but B and C are now in the top section. It looks like this test is mostly concerned with the existence of two traces, instead of caring about where the follow-up functions B and C are located, so I'm inclined to change the string to make the test pass.

Reviewed By: kennyyu

Differential Revision: D45408960

fbshipit-source-id: a7f4268c44d2d3c717422c1eda112cfc3df5d83f