Execute sub-commands smarter in test runner
Summary:
Don't use timeout, instead implement the timeout ourself. This lets us
check for failures quicker and better. For example, if the command
forks a child, the child will have the same stdout open. If the parent
crashes, the stdout will remain open because the child still has it
open. So, something like file_get_contents will block forever. If we
poll the process status as well, we can detect this and terminate
(instead of hanging).

Reviewed By: mdko

Differential Revision: D45434938

fbshipit-source-id: 3f5ea9390bd5282444eb692faf6de9f36507cb46