remove layers of indirection in naming_errors
Summary:
This diff is an RFC. I'd be grateful for your opinion on whether you think the new form of this test is easier to read.

I was struggling too hard to read the test. It had too many layers of indirection. I think my brain prefers to see a test just as a straightforward sequence of operations, with everything explicit.

Reviewed By: Wilfred

Differential Revision: D45419884

fbshipit-source-id: c6a09691c30dae144c042f8bfd66647342bf8fb0