Improve the readme
Summary:
Remove `sudo` from the command to install dependencies since it doesn't work with homebrew and not needed with D45371004.

Remove various noise such as horizontal rules and an out-of-place Meta logo and fix wording in a few places while at it.

Fixes https://github.com/facebook/fbthrift/issues/545 together with D45371004.

Reviewed By: prasad223

Differential Revision: D45446843

fbshipit-source-id: 7b5b57a891c489faac18627b883f9ac7de72a476