Refactor typedef to type structure conversion phase
Summary:
In order to only have a single entry point for type constant emitter, we have added bunch of crux to this function (is typedef, is opaque) where other than one caller, everything else always passes false.

In addition to this, going forwards, I'm planning to migrate some more of the type structure logic from HHVM to hackc to simplify implementation for case types. It will be helpful to have ast::typedef when we make decisions.

Reviewed By: paulbiss

Differential Revision: D45448774

fbshipit-source-id: 809021b3657e9687dfb2e6786c8e054745b3131a