Replace parsing_driver with a parse_ast function
Summary:
`parsing_driver` is not really a driver in the usual compiler sense but an implementation of a parser API that constructs and returns an AST. Now that its public API has been reduced to essentially one function, move the implementation detail class from the header to the source file replacing it in the header with a single function inventively called `parse_ast`. The name is shamelessly stolen from clang which has `ParseAST` performing nearly identical function. Rename the header and the source accordingly to make the purpose of the API clear.

This also makes the parser slightly easier to use since you no longer need to construct a "driver" that was never reused and likely couldn't be reused.

Before:
```
parsing_driver driver(sm, diags, filename, params);
auto programs = driver.parse();
```

After:
```
auto programs = parse_ast(sm, diags, filename, params);
```

#forcetdhashing

Reviewed By: avalonalex

Differential Revision: D45394853

fbshipit-source-id: b834e340d50ecb6cc26fb76ea696ec3624e68ef0