allow rename to shell out by symbol
Summary:
Internal

Like the last diff, we continue to aim to support unsaved files. In order to
remove a dependence on positions, I've added support to rename by symbol. This
required a few changes

1) I had to add some functions in ServerRename to take a Find_refs.action and
SymbolDefinition.t to short-circuit some of the lookup logic that
`go` goes. We already have both of those fields from our initial lookup to see
if it was a localvar or not.

2) I had to add a new mode for shelling out to take a Symbol directly, which is
the entire SymbolDefinition structure.

Since positions are returned by hh_server, they don't support edited files. I
intend to add this in the next diff, re-using the `go_for_single_file` logic I
implemented higher in the stack`

Reviewed By: ljw1004

Differential Revision: D44999555

fbshipit-source-id: 09819e6b422ad57499f993b18cfcc601ee1b1d80