let trait IsZeroInitializable fail union, member-object-pointer
Summary:
Closes: https://github.com/facebook/folly/issues/1271.

As reported in the GitHub issue, it is the intention to exclude types from the trait `IsZeroInitializable` if they are class or similar types. This similarity conceptually includes unions, but the implementation of the trait did not include union types. Fix.

As reported in a comment to the issue, whether member-object-pointer types are intrinsically zero-initialized is implementation-defined, and implementations differ.

Reviewed By: dmm-fb

Differential Revision: D42863011

fbshipit-source-id: 28fea02492f98642d7a3da620624425ee14b7a3b