report telemetry in CLIENT_CHECK_PARTIAL 1/2
Summary:
I'm working towards a metric for "client availability" T152127880. It will measure whether we were able to deliver results for "hh check", and how long they took.

It'd be useful, when looking at metrics for "hh check" performance, to know how long the underlying typecheck took to perform. I'll want to slice+dice, e.g. to narrow down performance only on those checks that involved large fanouts.

That's already provided in the CLIENT_CHECK event, both for streaming and non-streaming. In the case of streaming, it's present because ServerTypeCheck delivers a "Complete" sentinel at the end of the typecheck, and this sentinel includes the same telemetry that TYPE_CHECK_END provides.

But what if the user sees an error, which satisfies them, and they press Ctrl+C? In that case we produce CLIENT_CHECK_PARTIAL before ServerTypeCheck even completed, hence have no ability to slice-or-dice by fanout or distc. That's a shame because I think this audience will be particularly keen on Ctrl+C!

The answer I propose is that the errors-file can include "Telemetry" items as well as "Error" items during any stage of its typechecking logic. The client can accumulate those telemetry items.

This diff is merely preparatory. It invents a new variant type, which currently has only a single constructor, for the kinds of things that can appear in an errors-file:
```
type errors_file_item =
  | Errors of {
      errors: Errors.finalized_error list Relative_path.Map.t;
      timestamp: float;
    }
```

Reviewed By: CatherineGasnier

Differential Revision: D45468881

fbshipit-source-id: 0f2fda5a11c3365a64a378113cf0941e9002c33f