show callstack to stdout in case of uncaught exception
Summary:
In an earlier diff D45196752, I had split "CLIENT_BAD_EXIT" (which used to be emitted for all bad exits) into two:
* CLIENT_CHECK_BAD_EXIT for bad exits out of "hh check"
* CLIENT_BAD_EXIT for all other bad exits

The reason for this was to make it easier to write scuba queries that looked at the outcome of all "hh check" commands.

However, I missed something. In the case of uncaught exceptions, the CLIENT_BAD_EXIT path always used to print backtraces. This is nice so that if users encounter a problem, they can post a question in feedback groups and include the backtrace. I had accidentally forgotten backtraces for the new CLIENT_CHECK_BAD_EXIT. This diff restores those callstacks.

Reviewed By: CatherineGasnier

Differential Revision: D45471387

fbshipit-source-id: 9445c47f25d20b4557c4a15a42fac0b9af21a526