report telemetry in CLIENT_CHECK_PARTIAL 2/2
Summary:
The previous diff's summary already set the stage: we want ServerTypeCheck to be able to incrementally write telemetry into the errors-file, which a client is able to accumulate, and log in the cast of Ctrl+C CLIENT_CHECK_PARTIAL.

That's what this diff achieves. It adds a new kind of thing that the errors-file can include, a "telemetry" object:
```
type errors_file_item =
  | Errors of {
      errors: Errors.finalized_error list Relative_path.Map.t;
      timestamp: float;
    }
  | Telemetry of {
      key: string;
      telemetry: Telemetry.t;
    }
```

It changes clientCheckStatus.ml so it can accumulate those telemetry items, and report them in CLIENT_CHECK_PARTIAL. (also in CLIENT_CHECK).

It changes serverTypeCheck.ml to report telemetry into the streaming-errors file about the size of the fanout, and typing_check_service.ml to report telemetry into the file about whether distc was used.

Reviewed By: CatherineGasnier

Differential Revision: D45469371

fbshipit-source-id: 4209901f40b5794a56c5a463be600a64d5a5a2cd