log spinner even when hidden
Summary:
I recently added telemetry to log the state of the "hh_server is busy [...]" spinner, (1) at CLIENT_CHECK_BAD_EXIT, (2) every second in scuba table hh_client_spinner.

The telemetry mysteriously showed a lot of places where the spinner had "null". It also irritatingly showed that every time it got a SIGPIPE the spinner had "null".

The problem is that our code deliberately turns off the spinner temporarily (e.g. when it's about to write to stdout) and then turns it on again afterwards. We have no way of knowing whether a "null" spinner in the telemetry is there because no telemetry has yet been sent, or because it's being temporarily silenced.

This diff changes both places of telemetry so they log the most recent "Some" spinner, prefixing it with "[hidden]" if it's currently hidden.

Reviewed By: CatherineGasnier

Differential Revision: D45466420

fbshipit-source-id: 6ad2b0f3048de5cdf1021bdc8de8f6d82ea5696f