Canonicalize repo location (and use a std::fs:path for m_path instead of a string)
Summary: The Autoloader's `requestInit` method calls `getRepoRoot()` multiple times. This is slow because the path is stored as a `string`, then it's converted to a `std::filesystem::path`, and then it's canonicalized. Instead, just store the things as a `std::filesystem::path` to begin with and canonicalize the repo directory (and use the `dir()` method instead of using `path().parent_path()` for the repo root, since that what it is anyway)

Reviewed By: ricklavoie

Differential Revision: D45404359

fbshipit-source-id: 462d3a0877c9a091a284e248d17246e678be160b