Fix UB in shardalloc
Summary:
D44476382 enables building rust std library from source, which allows us to enable debug assertions in Rust. This caught an issue with undefined behaviour in Hack, where we were calling `ptr::write` with a misaligned pointer. With debugger attached, the bad address was aligned to alignment = 1, which is too small for `AllocHeader`.

This change proposes aligning the pointer to at *least* the alignment of AllocHeader, in the case where the provided `l` is too small.

Reviewed By: dtolnay

Differential Revision: D45539401

fbshipit-source-id: 51945d1f80dab24c6b9a583afe708282b2a3be8a