Use `AsyncFdSocket` when accepting Unix socket connections
Summary:
This is necessary to enable FD passing over Thrift. It's safe because...

(1) `Acceptor::makeNewAsyncSocket` is defined as
```
    return folly::AsyncSocket::UniquePtr(new folly::AsyncSocket(
        base, folly::NetworkSocket::fromFd(fd), 0, peerAddress));
```

That is, it is only capable of producing literal `AsyncSocket`s.

(2) `AsyncFdSocket` is a child of `AsyncSocket` that introduces a couple of new FD-handling methods, and bans the usage of the "send" and "read" ancillary data callbacks.
 - Adding new methods is safe.
 - The "read ancillary data" callback has no existing usage (besides in `AsyncFdSocket` itself).
 - The "send ancillary data" callback has a couple of uses, neither of which would interact with this change here.

Therefore, this change should be a fully transparent replacement for the `AsyncSocket` used to accept `AF_UNIX` connections today.

Reviewed By: AkramaMirza

Differential Revision: D43935120

fbshipit-source-id: 033d0a70b5b3b0b1d6de536525b1891ad480c2e8